refactor old structures
made struct s_vec3 => t_vec3 (union u_vec3): union of an array of 4
doubles and a struct of 4 doubles. should help the compiler vectorise
the math
added alias t_norm of t_vec3: should make it clearer if the variable is
expected to already be normalised or not
made t_line (struct s_line): struct of t_vec3 and t_norm, meant to make
line intersections easier to read.